added a nav bar and organized my files, also took of my picture but will be used later
