+++ src/App.js
@@ -1,8 +1,9 @@
 import React from 'react';
 //import logo from './joseBautista.jpeg';
 import './App.css';
-import Frontpage from './components/Frontpage.js'
+import Frontpage from './components/Front Page/Frontpage.js'
 
+import {BrowserRouter,Route,Switch, Redirect} from 'react-router-dom';
 
 
 const App = () => {
