--- src/components/Frontpage.css
@@ -1,50 +1,0 @@
-.App {
-    text-align: center;
-  }
-
-  
-
-  .Buttons{
-       height: 75px;
-    width: 250px;
-    font-size: 20px;
-  color: rgb(255, 230, 0);
-  background-color: rgb(0, 0, 0);
-  border-radius: 8px;
-  border-color: yellow;
-  opacity: .9;
-  
-  }
-  
-  .Front-pic {
-    height: 40vmin;
-    pointer-events: none;
-    opacity:.8 ;
-    border-radius: 8px;
-  }
-  
-  .App-header {
-    background: url('./IMG_7614 2.jpg');
-    min-height: 100vh;
-    background-position:center; 
-    
-  
-    background-repeat: no-repeat;
-    background-size:1440px 1000px;
-    display:flex;
-    flex-direction: column;
-    align-items: center;
-    justify-content: center;
-    font-size: calc(10px + 2vmin);
-    color: rgb(255, 230, 0);
-  }
-  
-  
-  .Website-Links{
-    color:rgb(255, 230, 0);
-    background-color: rgb(0, 0, 0);
-    border-radius: 8px;
-    align-content: stretch;
-  }
-  
-
